[org.colinux.devel @ 721c4fbd1e98f5e1f24c55e03353fb25d238306b] * Add message file to colinux-daemon.  The event logger prints only   one single line now, and no longer such stub text:
The description for Event ID ( 1 ) in Source ( coLinux ) cannot be
found. The local computer may not have the necessary registry
information or message DLL files to display messages from a remote
computer. You may be able to use the /AUXSOURCE= flag to retrieve this
description; see Help and Support for details. The following information
is part of the event: daemon: service control STOP.

* Messages compiled with mc.exe Message Compiler 1.12.2505 from SDK.
  Source and binary added to colinux source tree.

* Event ID changed from 1 to 0x40000001  (Severity=Informational, Code=1)

* Installer sets the path to message file into the registry.
  If you not use coLinux-installer, then please add it manualy with your
  path to coLinux installation. For example:

[HKEY_LOCAL_MACHINE\\SYSTEM\\CurrentControlSet\\Services\\Eventlog\\Application\\coLinux]
"EventMessageFile"="C:\\\\coLinux\\\\colinux-daemon.exe"


git-svn-id: http://svn.code.sf.net/p/colinux/code/branches/devel@598 688f6386-cf32-0410-9ead-a596f0484e40